Module Gobba.Util

module Dict : sig ... end

val take : int -> 'a list -> 'a list

Helper function to take the first elements of a list

val drop : int -> 'a list -> 'a list

Helper function to drop the first elements of a list

val last : 'a list -> 'a

val fst : ('a * 'b) -> 'a

val snd : ('a * 'b) -> 'b

val fst3 : ('a * 'b * 'c) -> 'a

val snd3 : ('a * 'b * 'c) -> 'b

val thd3 : ('a * 'b * 'c) -> 'c

val fstl : ('a * 'b) list -> 'a list

val sndl : ('a * 'b) list -> 'b list

val fst3l : ('a * 'b * 'c) list -> 'a list

val snd3l : ('a * 'b * 'c) list -> 'b list

val thd3l : ('a * 'b * 'c) list -> 'c list

val unzip : ('a * 'b) list -> 'a list * 'b list

Helper function to unzip a list of couples

val unzip3 : ('a * 'b * 'c) list -> 'a list * 'b list * 'c list

Helper function to unzip a list of triples

val zip : 'a list -> 'b list -> ('a * 'b) list

val zip3 : 'a list -> 'b list -> 'c list -> ('a * 'b * 'c) list