Module Gobba__Optimizer

val optimize : Gobba.Types.expr -> Gobba.Types.expr

Normalize an AST expression

val optimize_let : Gobba.Types.assignment_type list -> Gobba.Types.expr -> Gobba.Types.expr

val iterate_optimizer : Gobba.Types.expr -> Gobba.Types.expr

Apply the optimizer again and again on an expression until it is fully reduced and ready to be evaluated